+++ cpp/src/IceXML/Parser.cpp
@@ -394,7 +394,7 @@ IceXML::Parser::parse(istream& in, Handler& handler)
             {
                 isFinal = 1;
             }
-            if(XML_Parse(parser, buff, in.gcount(), isFinal) != 1)
+            if(XML_Parse(parser, buff, static_cast<int>(in.gcount()), isFinal) != 1)
             {
                 handler.error(XML_ErrorString(XML_GetErrorCode(parser)), XML_GetCurrentLineNumber(parser),
                               XML_GetCurrentColumnNumber(parser));
